OOP the theme
