+++ themes/10up-theme/assets/js/block-editor/block-editor-script.js
