+++ themes/10up-theme/blocks/example-block/markup.php
