+++ themes/10up-theme/composer.json
@@ -15,7 +15,7 @@
   },
   "autoload": {
     "psr-4": {
-      "TenUpTheme\\": "includes/classes/"
+      "TenUpTheme\\": "src/"
     }
   }
 }
