--- themes/10up-theme/includes/classes/.gitkeep
@@ -1,1 +1,0 @@
-# Basically just want to ignore the directory contents
