--- themes/10up-theme/includes/template-tags.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * Custom template tags for this theme.
- *
- * This file is for custom template tags only and it should not contain
- * functions that will be used for filtering or adding an action.
- *
- * All functions should be prefixed with TenUpTheme in order to prevent
- * pollution of the global namespace and potential conflicts with functions
- * from plugins.
- * Example: `tenup_theme_function()`
- *
- * @package TenUpTheme\Template_Tags
- *
- */
-
-// phpcs:ignoreFile
