+++ themes/10up-theme/package.json
@@ -30,7 +30,10 @@
       "editor-style-overrides": "./assets/js/admin/editor-style-overrides.js",
       "frontend": "./assets/js/frontend/frontend.js",
       "shared": "./assets/js/shared/shared.js",
-      "block-editor-script": "./includes/block-editor-script.js"
+      "block-editor-script": "./assets/js/block-editor/block-editor-script.js"
+    },
+    "paths": {
+      "blocksDir": "./blocks"
     }
   }
 }
